fix(http/verification): :bug: Si el usuario no existe, se devuelve un error directamente

+++ src/router/auth/update_user/validations/verify_code.js
@@ -4,6 +4,9 @@ import { ERRORS } from '../messages/errors.js'
 import { Verifications } from '../../../../database/model/verifications/verifications.js'
 
 export const verifyCode = async (code, dbUserData, res) => {
+	if (!dbUserData?.verification) {
+		return httpError(401, ERRORS.VERIFICATION_INVALID)(res)
+	}
 	const isValid = await compare(code, dbUserData.verification)
 	if (!isValid) return httpError(401, ERRORS.VERIFICATION_INVALID)(res)
 	await Verifications.deleteMany({ user_name: dbUserData.user_name })
